* Fix timeline dashboard2 (#39)

* Crud manager/duong (#41)

* CRUD-Manager/duong

* crud manager

* create post

* load back end

* hh

* hhh

* edit ui index

* Qa/phuong (#43)

* Fix timeline dashboard2

* delete QandA post

* Login/khoa (#44)

* fix dashboard

* style ui for dep hon

* Profile/chung (#45)

* fix Post Detail to display detail post

* filter and display trend

* authorization

* Accept incoming

* ađ account in department, my post

* updateDashboard

* no thing here

* Change commnet

* Profile/chung (#46)

* fix Post Detail to display detail post

* filter and display trend

* authorization

* Accept incoming

* ađ account in department, my post

* updateDashboard

* no thing here

* Change commnet

* fĩ change

* Login/khoa (#47)

* fix dashboard

* style ui for dep hon

* change color

* Profile/chung (#48)

* fix Post Detail to display detail post

* filter and display trend

* authorization

* Accept incoming

* ađ account in department, my post

* updateDashboard

* no thing here

* Change commnet

* fĩ change

* nothing

* change

* lastest change

* Profile/chung (#50)

* fix Post Detail to display detail post

* filter and display trend

* authorization

* Accept incoming

* ađ account in department, my post

* updateDashboard

* no thing here

* Change commnet

* fĩ change

* nothing

* change

* lastest change

* newest code

* new upload for Dashboard

---------

+++ Backend/routes/departmentRouter.js
@@ -1,6 +1,8 @@
 import express from "express";
 import Department from "../Model/departmentModel.js";
+import User from "../Model/userModel.js";
 import expressAsyncHandler from "express-async-handler";
+import { isAdmin, isAuth } from "../utils.js";
 
 const departmentRouter = express.Router();
 
@@ -25,16 +27,22 @@ departmentRouter.post(
     });
   })
 );
-
 departmentRouter.delete(
   "/:id",
+  isAuth,
   expressAsyncHandler(async (req, res) => {
-    const department = await Department.findById(req.params.id);
-    if (department) {
-      await department.deleteOne();
-      res.send({ message: "Department Deleted" });
-    } else {
-      res.status(404).send({ message: "Department Not Found" });
+    const users = await User.find({ department: req.params.id });
+    if(users.length === 0){
+      const department = await Department.findById(req.params.id);
+      if (department) {
+        await department.deleteOne();
+        res.send({ message: "Department Deleted" });
+      } else {
+        res.status(404).send({ message: "Department Not Found" });
+      }
+    }
+    else{
+      res.status(404).send({ message: "This department can be deleted" });
     }
   })
 );
